chore(configs): moved config files to down level directory
