+++ locales.js
@@ -1,5 +1,5 @@
-const rawLocales = require('../src/i18n/config.json');
-const defaultMessages = require('../src/i18n/locales/en.json');
+const rawLocales = require('./src/i18n/config.json');
+const defaultMessages = require('./src/i18n/locales/en.json');
 
 const locales = rawLocales.filter(locale => locale.enabled);
